Fixed tests for oauth and store
