+++ tests/application/shared.py
@@ -1,0 +1,11 @@
+from pydantic.dataclasses import dataclass
+
+
+@dataclass
+class MockHTTPResponse:
+    status_code: int
+    jsondata: dict
+    headers: dict = None  # type: ignore
+
+    def json(self):
+        return self.jsondata
